requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
schedule>=1.2.0
paramiko>=3.4.0
